b4038

+++ client/client.h
@@ -267,7 +267,7 @@ typedef struct client_static_s
 	float			frametime;			// seconds since last frame
 
 // screen rendering information
-	float			disable_screen;		// showing loading plaque between levels
+	unsigned		disable_screen;		// showing loading plaque between levels
 									// or changing rendering dlls
 									// if time gets > 30 seconds ahead, break it
 	int				disable_servercount;	// when we receive a frame and cl.servercount
